# contracts/openapi/minialalipay-api.yaml
openapi: 3.1.0
info:
  title: MiniAlalipay API
  version: 0.2.0
  description: MiniAlalipay 对外 HTTP 契约。业务操作只有在本文件中具备完整 Schema 后才允许实现。
servers:
  - url: http://localhost:8080
    description: 本地网关
...
paths:
...
  /api/v1/bank-cards:
    get:
      summary: 查询本人银行卡列表（仅已绑定，默认卡在前）
      operationId: listMyBankCards
      x-client-scope: C
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: 返回银行卡掩码列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BankCardListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      summary: 绑定银行卡（基于注册记录的绑卡流程，需先完成身份绑定和银行卡注册）
      description: 请求体包含注册时获得的完整卡号与三要素明文，服务端校验注册记录、三要素匹配及用户存储身份后完成绑定；重复绑定同一张卡返回 409。解绑后的卡其注册记录会释放回可绑定状态，可凭注册时的完整卡号与三要素重新绑定，每次重绑生成新的绑定记录。
      operationId: bindBankCard
      x-client-scope: C
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BindBankCardRequest' } } }
      responses:
        '200':
          description: 绑定成功，返回掩码卡片
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BankCardResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /api/v1/bank-cards/{cardId}:
    get:
      summary: 查询银行卡详情（全掩码字段）
      operationId: getBankCard
      x-client-scope: C
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - name: cardId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 26 }
      responses:
        '200':
          description: 返回掩码卡片详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BankCardResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: 解绑银行卡（软删为 UNBOUND 终态，同步释放注册记录支持重绑，解绑默认卡后递补最早活动卡）
      operationId: unbindBankCard
      x-client-scope: C
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - name: cardId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 26 }
      responses:
        '200':
          description: 解绑成功，对应注册记录已释放回 REGISTERED，该卡可重新绑定
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/bank-cards/{cardId}/balance:
    get:
      summary: 查询银行卡虚拟余额
      operationId: getBankCardBalance
      x-client-scope: C
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - name: cardId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 26 }
      responses:
        '200':
          description: 返回银行卡虚拟余额（分）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        type: object
                        required: [balanceFen]
                        properties:
                          balanceFen: { type: integer, format: int64, description: 虚拟余额（分） }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/bank-cards/{cardId}/recharge:
    post:
      summary: 银行卡充值（银行卡给账户充钱，虚拟余额 → 账户余额）
      description: |
        通过 Seata TCC 全局事务将银行卡虚拟余额扣减并增加到账户余额（银行卡给账户充钱）。单笔 0.01–50000.00 元，日累计充值 50000.00 元，且不超过银行卡当前虚拟余额。
        paymentPassword 属敏感字段，只允许出现在请求体，不得进入日志、URL、浏览器存储或埋点。
      operationId: rechargeBankCard
      x-client-scope: C
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: cardId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 26 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [amountFen, paymentPassword]
              properties:
                amountFen: { type: integer, format: int64, minimum: 1, maximum: 500000000, description: 充值金额（分），单笔 1–500000000 } 
                paymentPassword: { type: string, pattern: '^\d{6}$', writeOnly: true, description: 六位数字支付密码 }
      responses:
        '200':
          description: 充值成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/bank-cards/{cardId}/withdraw:
    post:
      summary: 银行卡提现（账户给银行卡充钱，账户余额 → 虚拟余额）
      description: |
        通过 Seata TCC 全局事务将账户余额提现到银行卡虚拟余额（账户给银行卡充钱）。单笔 0.01–50000.00 元，日累计提现 50000.00 元。
        paymentPassword 属敏感字段，只允许出现在请求体，不得进入日志、URL、浏览器存储或埋点。
      operationId: withdrawBankCard
      x-client-scope: C
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: cardId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 26 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [amountFen, paymentPassword]
              properties:
                amountFen: { type: integer, format: int64, minimum: 1, maximum: 500000000, description: 提现金额（分），单笔 1–500000000 } 
                paymentPassword: { type: string, pattern: '^\d{6}$', writeOnly: true, description: 六位数字支付密码 }
      responses:
        '200':
          description: 提现成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /api/v1/recharges/{id}:
    get:
      summary: 查询本人模拟充值来源订单
      operationId: getRecharge
      x-client-scope: C
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200': { description: 返回来源订单状态, content: { application/json: { schema: { $ref: '#/components/schemas/RechargeOrderResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 会话令牌
      description: 由网关验证的会话令牌，不得写入日志或 URL。
    cookieAuth:
      type: apiKey
      in: cookie
      name: MINIALALIPAY_SESSION
      description: HttpOnly、Secure、SameSite Cookie；写请求同时校验 CSRF 令牌。
    serviceAuth:
      type: apiKey
      in: header
      name: X-Service-Token
      description: 仅用于服务间鉴权的内部令牌，由部署环境注入，不得写入日志或经过前端网关。
    internalServiceAuth:
      type: apiKey
      in: header
      name: X-Internal-Service-Token
      description: 账户中心既有内部资金接口使用的服务身份令牌，不得经过前端网关或写入日志。
  parameters:
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      description: 客户端请求编号；缺失或不合法时由网关生成。
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: 写操作幂等键，同一主体和接口作用域至少保留 24 小时。
      schema:
        type: string
        minLength: 16
        maxLength: 64
        pattern: '^[A-Za-z0-9._:-]+$'
...
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 26 }
  schemas:
...
    ApiResponse:
      type: object
      additionalProperties: false
      required: [code, message, requestId, traceId, data]
      properties:
        code:
          type: string
          description: 稳定的机器可读结果码。
          example: OK
        message:
          type: string
          description: 面向调用方的中文说明。
          example: 成功
        requestId:
          type: string
          description: 一次网关请求的编号。
          example: req_01K1ABCDEF2GH3JK4MN5PQRSTV
        traceId:
          type: [string, 'null']
          description: 跨服务链路编号；链路尚未建立时为空。
          example: 6f8b0f0a0ad64fd1b9eb5be13e146a82
        data:
          description: 业务数据；错误响应只允许包含安全、可操作的冲突信息。
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: [object, 'null']
              additionalProperties: true
...
    BindBankCardRequest:
      type: object
      required: [cardNumber, holderName, idCard, phone]
      properties:
        cardNumber: { type: string, minLength: 16, maxLength: 26, description: 完整银行卡号，16 至 19 位数字，允许空格分组；仅出现一次，不落存储 }
        holderName: { type: string, minLength: 1, maxLength: 32, description: 持卡人姓名明文 }
        idCard: { type: string, minLength: 18, maxLength: 18, description: 身份证号，18 位，末位可为 X/x }
        phone: { type: string, minLength: 11, maxLength: 11, description: 预留手机号，11 位 }
    BankCard:
      type: object
      required: [cardId, bankCode, bankName, cardType, cardLast4, holderMasked, idCardMasked, phoneMasked, balanceFen, isDefault, status, boundAt]
      properties:
        cardId: { type: string, description: 银行卡 ID，26 位字符串 }
        bankCode: { type: string, description: 银行编码，如 ICBC、CMB }
        bankName: { type: string, description: 银行名称 }
        cardType: { type: string, enum: [DEBIT, CREDIT], description: DEBIT 借记卡，CREDIT 信用卡 }
        cardLast4: { type: string, description: 卡号后 4 位，前端拼接展示为 **** **** **** 1234 }
        holderMasked: { type: string, description: 持卡人姓名掩码，如 张*三 }
        idCardMasked: { type: string, description: 身份证号掩码 }
        phoneMasked: { type: string, description: 预留手机号掩码，如 138****5678 }
        balanceFen: { type: integer, format: int64, description: 虚拟余额（分），充值增加、提现/支付扣减，与账户余额独立 }
        isDefault: { type: boolean, description: 是否默认卡，同一用户至多一张 }
        status: { type: string, enum: [ACTIVE, UNBOUND], description: ACTIVE 已绑定，UNBOUND 已解绑（终态） }
        boundAt: { type: string, format: date-time, description: 绑定时间 }
    BankCardListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/BankCard'
    BankCardResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/BankCard'
...
    RechargeOrderResponse:
      type: object
      additionalProperties: false
      required: [rechargeOrderId, amountFen, status, version, createdAt]
      properties:
        rechargeOrderId: { type: string, description: 充值来源订单 ID。 }
        amountFen: { type: integer, format: int64, description: 充值金额，单位为分。 }
        status:
          type: string
          description: 来源订单状态；PENDING_CHANNEL 限额已预占等待渠道结果、PROCESSING 已受理、SUCCESS 已由统一资金交易终态投影、REJECTED 渠道拒绝并释放日额度、CANCELLED 已取消、MANUAL_REVIEW 人工复核。PROCESSING 与未知结果绝不表示成功。
        version: { type: integer, format: int64, description: 订单版本，用于 CAS。 }
        createdAt: { type: string, format: date-time, description: 创建时间。 }
...
  responses:
    BadRequest:
      description: 请求格式或字段不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证或会话失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: 幂等、版本、状态或唯一约束冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在或调用方无权感知
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnprocessableEntity:
      description: 可理解的业务规则拒绝
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: 请求超过频率或安全尝试限制
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 未分类内部错误，不返回内部细节
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: TCC、模型或关键依赖暂不可用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - bearerAuth: []
  - cookieAuth: []